+++ NEWS.rst
@@ -4,6 +4,8 @@ Version History
 0.8.0 (not yet released)
 ------------------------
 
+* multi_decompress_into_buffer() API for decompressing multiple frames as a
+  single operation, as effeciently as possible.
 * Support for multi-threaded compression.
 * Disable deprecation warnings when compiling CFFI module.
 * Fixed memory leak in train_dictionary().